backend/requirements.txt
# FastAPI and ASGI
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.10.0

# Data validation
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0

# Environment
python-dotenv>=1.0.0

# Development
pytest>=7.0.0
httpx>=0.24.0

# PDF Generation
reportlab>=4.0.0
Pillow>=10.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
